fix: ChunkConfig truncated entropy_packing_iterations (u8→u32) and chain_id (u16→u64)
Pre-existing bug: ChunkConfig.entropy_packing_iterations was u8, truncating
the testing config value of 1000 to 232. This caused irys_packing::unpack
to compute wrong entropy, producing garbage 'unpacked' data.

Also widened chain_id from u16 to u64 to match ConsensusConfig.

With this fix, 3/4 PD chunk P2P pull integration tests pass:
- test_pd_chunk_p2p_happy_path: PASS
- test_pd_chunk_p2p_multiple_txs: PASS
- test_pd_chunk_p2p_mempool_path: PASS
- test_pd_chunk_p2p_deduplication: FAIL (nonce issue in test design)

+++ crates/domain/src/models/chunk_provider.rs
@@ -172,8 +172,8 @@ impl irys_types::chunk_provider::ChunkStorageProvider for ChunkProvider {
         irys_types::chunk_provider::ChunkConfig {
             num_chunks_in_partition: self.config.consensus.num_chunks_in_partition,
             chunk_size: self.config.consensus.chunk_size,
-            entropy_packing_iterations: self.config.consensus.entropy_packing_iterations as u8,
-            chain_id: self.config.consensus.chain_id as u16,
+            entropy_packing_iterations: self.config.consensus.entropy_packing_iterations,
+            chain_id: self.config.consensus.chain_id,
         }
     }
 }
